Fix compiler warnings in wmbus sources
- Telegram.cpp: the ELL V branch had a bare string statement; log it
  with warning() as intended.
- meters.cpp: time_t is long long on ESP8266, format it with %lld.
- driver_hydrus.cpp: mark the constructor inline so GCC no longer
  reports it as defined but unused (it is only referenced from the
  registerDriver lambda).

+++ components/wmbus/meters.cpp
@@ -579,7 +579,7 @@ string MeterCommonImplementation::unixTimestampOfUpdate()
 {
     char ut[40];
     memset(ut, 0, sizeof(ut));
-    snprintf(ut, sizeof(ut) - 1, "%lu", datetime_of_update_);
+    snprintf(ut, sizeof(ut) - 1, "%lld", (long long) datetime_of_update_);
     return string(ut);
 }
